erfgerg

+++ main.py
@@ -20,7 +20,7 @@ async def on_message(message):
       
       elif 0<=len(message.content)<=8:
           await message.delete()
-  elif  message.channel.id ==999823889653846060:
+  elif  message.channel.id ==999823889653846060 and len(message.content)!="suggest":
     await message.delete()
